Support optional usage of gRPC.
