+++ CMakeLists.txt
@@ -3,34 +3,64 @@ enable_testing()
 
 project(lightstep-tracer)
 
+# ==============================================================================
+# Version information
+
+set(LIGHTSTEP_VERSION_MAJOR "0")
+set(LIGHTSTEP_VERSION_MINOR "4")
+set(LIGHTSTEP_VERSION_PATCH "0")
+set(LIGHTSTEP_VERSION_STRING
+  "${LIGHTSTEP_VERSION_MAJOR}.${LIGHTSTEP_VERSION_MINOR}.${LIGHTSTEP_VERSION_PATCH}")
+
+# ==============================================================================
+# Set up cpack
+
+set(CPACK_PACKAGE_DESCRIPTION_SUMMARY 
+                  "A LightStep implementation of the C++ OpenTracing API")
+set(CPACK_PACKAGE_VENDOR "lightstep.com")
+set(CPACK_PACKAGE_DESCRIPTION_FILE "${CMAKE_CURRENT_SOURCE_DIR}/README.md")
+set(CPACK_RESOURCE_FILE_LICENSE "${CMAKE_CURRENT_SOURCE_DIR}/LICENSE")
+
+set(CPACK_PACKAGE_VERSION_MAJOR ${LIGHTSTEP_VERSION_MAJOR})
+set(CPACK_PACKAGE_VERSION_MINOR ${LIGHTSTEP_VERSION_MINOR})
+set(CPACK_PACKAGE_VERSION_PATCH ${LIGHTSTEP_VERSION_PATCH})
+include(cpack)
+
+
+# ==============================================================================
+# Set up options
+
 option(LS_WITH_ASAN "Generate tests using address-sanitizer." OFF)
 option(LS_WITH_TSAN "Generate tests using thread-sanitizer." OFF)
 option(LS_WITH_GRPC "Build with support for gRPC." ON)
 
-SET(CPACK_PACKAGE_DESCRIPTION_SUMMARY 
-                  "A LightStep implementation of the C++ OpenTracing API")
-SET(CPACK_PACKAGE_VENDOR "lightstep.com")
-SET(CPACK_PACKAGE_DESCRIPTION_FILE "${CMAKE_CURRENT_SOURCE_DIR}/README.md")
-SET(CPACK_RESOURCE_FILE_LICENSE "${CMAKE_CURRENT_SOURCE_DIR}/LICENSE")
 
-SET(CPACK_PACKAGE_VERSION_MAJOR "0")
-SET(CPACK_PACKAGE_VERSION_MINOR "4")
-SET(CPACK_PACKAGE_VERSION_PATCH "0")
-include(cpack)
+# ==============================================================================
+# Find packages
 
-find_package(PkgConfig REQUIRED)
 find_package(Protobuf REQUIRED)
-
-pkg_search_module(GRPC REQUIRED grpc)
-pkg_search_module(GRPCPP REQUIRED grpc++)
 
 find_path(OPENTRACING_INCLUDE_DIR NAMES opentracing/tracer.h)
 find_library(OPENTRACING_LIB opentracing)
 
+include_directories(SYSTEM ${OPENTRACING_INCLUDE_DIR})
+include_directories(SYSTEM ${PROTOBUF_INCLUDE_DIRS})
+
 set(LIGHTSTEP_LINK_LIBRARIES ${OPENTRACING_LIB}
-                             ${PROTOBUF_LIBRARIES} 
-                             ${GRPC_LIBRARIES}
-                             ${GRPCPP_LIBRARIES})
+                             ${PROTOBUF_LIBRARIES})
+
+if (LS_WITH_GRPC)                           
+  find_package(PkgConfig REQUIRED)
+  pkg_search_module(GRPC REQUIRED grpc)
+  pkg_search_module(GRPCPP REQUIRED grpc++)
+  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}  ${GRPC_CFLAGS_OTHER} \
+                       ${GRPCPP_CFLAGS_OTHER}")
+  list(APPEND LIGHTSTEP_LINK_LIBRARIES ${GRPC_LIBRARIES} ${GRPCPP_LIBRARIES})  
+  include_directories(SYSTEM ${GRPC_INCLUDE_DIRS} ${GRPCPP_INCLUDE_DIRS})
+endif()
+
+# ==============================================================================
+# Configure sanitizers
 
 set(WITH_ASAN OFF)
 if (LS_WITH_ASAN AND (("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang") OR
@@ -49,13 +79,5 @@ if (LS_WITH_TSAN AND (("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang") OR
    set(TSAN_CXX_FLAGS -O1 -g -fsanitize=thread)
    set(TSAN_LD_FLAGS -fsanitize=thread)
 endif()
-
-
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}  ${GRPC_CFLAGS_OTHER} \
-                     ${GRPCPP_CFLAGS_OTHER}")
-
-include_directories(SYSTEM ${GRPC_INCLUDE_DIRS} ${GRPCPP_INCLUDE_DIRS})
-include_directories(SYSTEM ${OPENTRACING_INCLUDE_DIR})
-include_directories(SYSTEM ${PROTOBUF_INCLUDE_DIRS})
 
 add_subdirectory(c++11)
