+++ c++11/example/tutorial/CMakeLists.txt
@@ -1,3 +1,4 @@
-add_executable(tutorial tutorial.cpp)
-target_link_libraries(tutorial lightstep_tracer ${OPENTRACING_LIB} ${PROTOBUF_LIBRARIES})
-
+if (LS_WITH_GRPC)
+  add_executable(tutorial tutorial.cpp)
+  target_link_libraries(tutorial lightstep_tracer ${LIGHTSTEP_LINK_LIBRARIES})
+endif()
