+++ c++11/version.h.in
@@ -1,5 +1,3 @@
 #pragma once
 
-#define LIGHTSTEP_VERSION "${CPACK_PACKAGE_VERSION_MAJOR}." \
-                          "${CPACK_PACKAGE_VERSION_MINOR}." \
-                          "${CPACK_PACKAGE_VERSION_PATCH}"
+#define LIGHTSTEP_VERSION "${LIGHTSTEP_VERSION_STRING}"
